refactor: cleaned up service directory structure
