+++ BigBaseV2/src/gui.cpp
@@ -1,22 +1,13 @@
 #include "common.hpp"
-#include "fiber_pool.hpp"
 #include "gta/player.hpp"
-#include "gta_util.hpp"
 #include "gui.hpp"
-#include "logger.hpp"
-#include "memory/module.hpp"
-#include "memory/pattern.hpp"
 #include "natives.hpp"
-#include "pointers.hpp"
-#include "renderer.hpp"
 #include "script.hpp"
-
 
 #include <imgui.h>
 #include "widgets/imgui_hotkey.hpp"
 
 #include "views/view.hpp"
-#include "services/notification_service.hpp"
 
 namespace big
 {
