+++ BigBaseV2/src/gui/handling/handling_search.cpp
@@ -3,7 +3,7 @@
 #include "natives.hpp"
 #include "script.hpp"
 #include "thread_pool.hpp"
-#include "services/vehicle_service.hpp"
+#include "services/vehicle/vehicle_service.hpp"
 #include "views/view.hpp"
 
 namespace big
