+++ BigBaseV2/src/gui/handling/handling_tabs.hpp
@@ -1,6 +1,6 @@
 #pragma once
 #include "current_profile/current_profile_tabs.hpp"
-#include "services/vehicle_service.hpp"
+#include "services/vehicle/vehicle_service.hpp"
 
 namespace big::tab_handling
 {
