+++ BigBaseV2/src/hooks/network_player_mgr.cpp
@@ -1,5 +1,5 @@
 #include "hooking.hpp"
-#include "services/player_service.hpp"
+#include "services/players/player_service.hpp"
 
 namespace big
 {
