+++ BigBaseV2/src/services/globals/globals_service.cpp
