+++ BigBaseV2/src/services/vehicle_helper/vehicle_helper.hpp
