+++ BigBaseV2/src/views/core/view_active_view.cpp
@@ -1,5 +1,6 @@
 #include "views/view.hpp"
-#include "services/gui_service.hpp"
+#include "pointers.hpp"
+#include "services/gui/gui_service.hpp"
 
 namespace big
 {
