+++ BigBaseV2/src/views/core/view_notifications.cpp
@@ -1,5 +1,5 @@
+#include "pointers.hpp"
 #include "views/view.hpp"
-#include "services/notification_service.hpp"
 
 namespace big
 {
