+++ BigBaseV2/src/views/players/view_player.cpp
@@ -1,10 +1,10 @@
-#include "views/view.hpp"
-#include "services/player_service.hpp"
 #include "gta_util.hpp"
-#include "util/misc.hpp"
+#include "services/players/player_service.hpp"
 #include "util/globals.hpp"
+#include "util/misc.hpp"
 #include "util/ped.hpp"
 #include "util/teleport.hpp"
+#include "views/view.hpp"
 
 namespace big
 {
