+++ BigBaseV2/src/views/players/view_players.cpp
@@ -1,5 +1,7 @@
+#include "pointers.hpp"
+#include "services/gui/gui_service.hpp"
+#include "services/players/player_service.hpp"
 #include "views/view.hpp"
-#include "services/gui_service.hpp"
 
 namespace big
 {
