+++ BigBaseV2/src/views/self/view_mobile.cpp
@@ -1,7 +1,7 @@
-#include "views/view.hpp"
 #include "fiber_pool.hpp"
+#include "services/mobile/mobile_service.hpp"
 #include "util/mobile.hpp"
-#include "services/mobile_service.hpp"
+#include "views/view.hpp"
 
 namespace big
 {
