+++ BigBaseV2/src/views/self/view_weapons.cpp
@@ -1,7 +1,7 @@
 #include "core/data/custom_weapons.hpp"
 #include "fiber_pool.hpp"
+#include "natives.hpp"
 #include "gta/Weapons.h"
-#include "script.hpp"
 #include "core/data/special_ammo_types.hpp"
 #include "views/view.hpp"
