+++ BigBaseV2/src/views/settings/view_settings.cpp
@@ -1,7 +1,6 @@
 #include "views/view.hpp"
 #include "widgets/imgui_hotkey.hpp"
 #include "script_mgr.hpp"
-#include "services/notification_service.hpp"
 
 namespace big
 {
