+++ BigBaseV2/src/views/vehicle/view_lsc.cpp
@@ -1,6 +1,7 @@
 #include "fiber_pool.hpp"
 #include "natives.hpp"
-#include "services/vehicle_helper.h"
+#include "script.hpp"
+#include "services/vehicle_helper/vehicle_helper.hpp"
 #include "views/view.hpp"
 
 namespace big
