+++ BigBaseV2/src/views/vehicle/view_spawn.cpp
@@ -1,7 +1,7 @@
 #include "views/view.hpp"
 #include "fiber_pool.hpp"
 #include "natives.hpp"
-#include "services/vehicle_preview_service.hpp"
+#include "services/vehicle_preview/vehicle_preview_service.hpp"
 #include "util/vehicle.hpp"
 
 namespace big
